src/main/java/com/github/brunomndantas/jscrapper/support/parser/collection/attribute/reference/CollectionDoubleAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.collection.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.collection.attribute.CollectionAttributeParser;
import org.openqa.selenium.WebElement;

public class CollectionDoubleAttributeParser extends CollectionAttributeParser {

    protected CollectionDoubleAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Double.parseDouble(element.getAttribute(super.attribute));
    }

}
